exam-admin/requirements.txt
# my deps
google-cloud-firestore
cryptography
google-auth
sendgrid
